Refactor body selectors
