+++ src/utils.ts
@@ -1,9 +1,6 @@
-import { OutgoingHttpHeaders, ServerResponse, IncomingMessage } from 'http'
+import { OutgoingHttpHeaders, ServerResponse } from 'http'
 import { ResponseObject, Selectors, Context } from './types'
 import { Stream } from 'stream'
-import { createError } from './error'
-import contentType from 'content-type'
-import getRawBody from 'raw-body'
 
 /**
  * Factory function for creating http responses
@@ -163,9 +160,6 @@ export function resolveSelectors<A extends any[]>(
   return Promise.all(selectors.map(selector => selector(context))) as Promise<A>
 }
 
-const { NODE_ENV } = process.env
-const DEV = NODE_ENV === 'development'
-
 function isStream(stream: any) {
   return (
     stream !== null &&
@@ -213,18 +207,7 @@ export function send(res: ServerResponse, code: number = 200, obj: any = null) {
   let str = obj
 
   if (typeof obj === 'object' || typeof obj === 'number') {
-    // We stringify before setting the header
-    // in case `JSON.stringify` throws and a
-    // 500 has to be sent instead
-
-    // the `JSON.stringify` call is split into
-    // two cases as `JSON.stringify` is optimized
-    // in V8 if called with only one argument
-    if (DEV) {
-      str = JSON.stringify(obj, null, 2)
-    } else {
-      str = JSON.stringify(obj)
-    }
+    str = JSON.stringify(obj)
 
     if (!res.getHeader('Content-Type')) {
       res.setHeader('Content-Type', 'application/json; charset=utf-8')
@@ -233,69 +216,4 @@ export function send(res: ServerResponse, code: number = 200, obj: any = null) {
 
   res.setHeader('Content-Length', Buffer.byteLength(str))
   res.end(str)
-}
-
-export function parseJSON(str: string) {
-  try {
-    return JSON.parse(str)
-  } catch (err) {
-    throw createError(400, 'Invalid JSON', err)
-  }
-}
-
-interface BodyParserOptions {
-  encoding?: string
-  limit?: string | number
-}
-
-// Maps requests to buffered raw bodies so that
-// multiple calls to `json` work as expected
-const rawBodyMap = new WeakMap()
-
-export async function buffer(
-  req: IncomingMessage,
-  { encoding, limit }: BodyParserOptions = {}
-) {
-  const type = req.headers['content-type'] || 'text/plain'
-  const length = req.headers['content-length']
-
-  // eslint-disable-next-line no-undefined
-  if (encoding === undefined) {
-    encoding = contentType.parse(type).parameters.charset
-  }
-
-  const body = rawBodyMap.get(req)
-
-  if (body) {
-    return body
-  }
-
-  if (req.complete) {
-    throw createError(500, `Body has been parsed already`)
-  }
-
-  let buf
-  try {
-    buf = await getRawBody(req, { limit, length, encoding })
-  } catch (error) {
-    if (error.type === 'entity.too.large') {
-      throw createError(413, `Body exceeded ${limit} limit`, error)
-    } else {
-      throw createError(400, 'Invalid body', error)
-    }
-  }
-  rawBodyMap.set(req, buf)
-
-  return buf
-}
-
-export function text(
-  req: IncomingMessage,
-  { limit, encoding }: BodyParserOptions = {}
-) {
-  return buffer(req, { limit, encoding }).then(body => body.toString(encoding))
-}
-
-export function json(req: IncomingMessage, opts?: BodyParserOptions) {
-  return text(req, opts).then(body => parseJSON(body))
 }
